// src/services/api/generated/models/board.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BaudRateEnum } from './baud-rate-enum';
// May contain unused imports in some cases
// @ts-ignore
import type { BoardCapability } from './board-capability';
// May contain unused imports in some cases
// @ts-ignore
import type { BoardStatusEnum } from './board-status-enum';
// May contain unused imports in some cases
// @ts-ignore
import type { DataBitsEnum } from './data-bits-enum';
// May contain unused imports in some cases
// @ts-ignore
import type { ParityEnum } from './parity-enum';
// May contain unused imports in some cases
// @ts-ignore
import type { StopBitsEnum } from './stop-bits-enum';

export interface Board {
    'id': number;
    'name': string;
    'serial_number': string;
    'description'?: string;
    'uart_port': string;
    'baud_rate'?: BaudRateEnum;
    'data_bits'?: DataBitsEnum;
    'stop_bits'?: StopBitsEnum;
    'parity'?: ParityEnum;
    'status': BoardStatusEnum;
    'last_seen_at': string | null;
    'capabilities': Array<BoardCapability>;
}